refactor(api): rename repository file
