+++ api/scripts/organization/update-organizations-identity-provider-for-campaigns.js
@@ -6,7 +6,7 @@ import _ from 'lodash';
 import { disconnect } from '../../db/knex-database-connection.js';
 import { updateOrganizationIdentityProviderForCampaigns } from '../../lib/domain/usecases/update-organization-identity-provider-for-campaigns.js';
 import { DomainTransaction } from '../../lib/infrastructure/DomainTransaction.js';
-import * as organizationForAdminRepository from '../../src/organizational-entities/infrastructure/repositories/organization-for-admin-repository.js';
+import * as organizationForAdminRepository from '../../src/organizational-entities/infrastructure/repositories/organization-for-admin.repository.js';
 import { checkCsvHeader, parseCsvWithHeader } from '../helpers/csvHelpers.js';
 
 const modulePath = fileURLToPath(import.meta.url);
