+++ api/src/organizational-entities/domain/usecases/index.js
@@ -4,13 +4,13 @@ import { fileURLToPath } from 'node:url';
 import { injectDependencies } from '../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
 import * as organizationFeatureRepository from '../../infrastructure/repositories/organization-feature-repository.js';
-import * as organizationForAdminRepository from '../../infrastructure/repositories/organization-for-admin-repository.js';
+import * as organizationForAdminRepository from '../../infrastructure/repositories/organization-for-admin.repository.js';
 
 const path = dirname(fileURLToPath(import.meta.url));
 
 /**
  * @typedef {import ('../../infrastructure/repositories/organization-feature-repository.js')} OrganizationFeatureRepository
- * @typedef {import ('../../infrastructure/repositories/organization-for-admin-repository.js')} OrganizationForAdminRepository
+ * @typedef {import ('../../infrastructure/repositories/organization-for-admin.repository.js')} OrganizationForAdminRepository
  */
 
 const repositories = {
