+++ api/tests/organizational-entities/integration/infrastructure/repositories/organization-for-admin.repository.test.js
@@ -3,7 +3,7 @@ import _ from 'lodash';
 import { MissingAttributesError, NotFoundError } from '../../../../../lib/domain/errors.js';
 import { OrganizationInvitation } from '../../../../../lib/domain/models/OrganizationInvitation.js';
 import { OrganizationForAdmin } from '../../../../../src/organizational-entities/domain/models/OrganizationForAdmin.js';
-import * as organizationForAdminRepository from '../../../../../src/organizational-entities/infrastructure/repositories/organization-for-admin-repository.js';
+import * as organizationForAdminRepository from '../../../../../src/organizational-entities/infrastructure/repositories/organization-for-admin.repository.js';
 import { ORGANIZATION_FEATURE } from '../../../../../src/shared/domain/constants.js';
 import { catchErr, databaseBuilder, domainBuilder, expect, knex, sinon } from '../../../../test-helper.js';
